Split propagation remote control module
