+++ crates/apps/reticulumd/src/bin/reticulumd/bridge_remote_control_link.rs
@@ -1,0 +1,249 @@
+use super::*;
+
+pub(super) async fn resolve_remote_identity(
+    transport: &Transport,
+    remote_hash: &AddressHash,
+    timeout: Duration,
+) -> Result<Option<Identity>, std::io::Error> {
+    transport.request_path(remote_hash, None, None).await;
+    let deadline = tokio::time::Instant::now() + timeout.min(Duration::from_secs(12));
+    let mut remote_identity = transport.destination_identity(remote_hash).await;
+
+    while remote_identity.is_none() && tokio::time::Instant::now() < deadline {
+        tokio::time::sleep(Duration::from_millis(250)).await;
+        remote_identity = transport.destination_identity(remote_hash).await;
+    }
+
+    if remote_identity.is_some() {
+        tokio::time::sleep(Duration::from_millis(500)).await;
+    }
+
+    Ok(remote_identity)
+}
+
+pub(super) async fn open_refreshed_remote_link(
+    transport: &Transport,
+    remote_hash: &AddressHash,
+    destination: DestinationDesc,
+    timeout: Duration,
+) -> Result<Arc<tokio::sync::Mutex<Link>>, std::io::Error> {
+    let link = transport.link(destination).await;
+    match await_link_activation(transport, &link, timeout).await {
+        Ok(()) => Ok(link),
+        Err(err) if err.kind() == std::io::ErrorKind::TimedOut => {
+            link.lock().await.close();
+            transport.request_path(remote_hash, None, None).await;
+            tokio::time::sleep(Duration::from_secs(1)).await;
+            let retry_link = transport.link(destination).await;
+            await_link_activation(transport, &retry_link, timeout).await.map_err(|retry_err| {
+                std::io::Error::new(
+                    retry_err.kind(),
+                    format!(
+                        "{retry_err}; retried after refreshing propagation node path due to {err}"
+                    ),
+                )
+            })?;
+            Ok(retry_link)
+        }
+        Err(err) => Err(err),
+    }
+}
+
+pub(super) fn build_link_identify_payload(
+    identity: &PrivateIdentity,
+    link_id: &AddressHash,
+) -> Vec<u8> {
+    let mut public_key = Vec::with_capacity(64);
+    public_key.extend_from_slice(identity.as_identity().public_key.as_bytes());
+    public_key.extend_from_slice(identity.as_identity().verifying_key.as_bytes());
+
+    let mut signed_data = Vec::with_capacity(16 + public_key.len());
+    signed_data.extend_from_slice(link_id.as_slice());
+    signed_data.extend_from_slice(public_key.as_slice());
+    let signature = identity.sign(signed_data.as_slice());
+
+    let mut payload = Vec::with_capacity(public_key.len() + signature.to_bytes().len());
+    payload.extend_from_slice(public_key.as_slice());
+    payload.extend_from_slice(signature.to_bytes().as_slice());
+    payload
+}
+
+pub(super) fn build_link_request_payload(
+    path: &str,
+    data: rmpv::Value,
+) -> Result<Vec<u8>, std::io::Error> {
+    let timestamp = SystemTime::now().duration_since(UNIX_EPOCH).unwrap_or_default().as_secs_f64();
+    let path_hash = address_hash(path.as_bytes());
+    rmp_serde::to_vec(&rmpv::Value::Array(vec![
+        rmpv::Value::F64(timestamp),
+        rmpv::Value::Binary(path_hash.to_vec()),
+        data,
+    ]))
+    .map_err(|err| std::io::Error::new(std::io::ErrorKind::InvalidData, err))
+}
+
+pub(super) async fn send_link_context_packet(
+    transport: &Transport,
+    link: &Arc<tokio::sync::Mutex<Link>>,
+    context: PacketContext,
+    payload: &[u8],
+) -> Result<Option<[u8; 16]>, std::io::Error> {
+    let (packet, ingress_iface) = {
+        let guard = link.lock().await;
+        if guard.status() != LinkStatus::Active {
+            return Err(std::io::Error::new(
+                std::io::ErrorKind::TimedOut,
+                "propagation control link is not active",
+            ));
+        }
+
+        let Some(ingress_iface) = guard.ingress_iface() else {
+            return Err(std::io::Error::new(
+                std::io::ErrorKind::NotConnected,
+                "propagation control link has no bound interface",
+            ));
+        };
+        let mut packet_data = PacketDataBuffer::new();
+        let cipher_len = {
+            let ciphertext = guard
+                .encrypt(payload, packet_data.accuire_buf_max())
+                .map_err(|_| std::io::Error::other("failed to encrypt link packet"))?;
+            ciphertext.len()
+        };
+        packet_data.resize(cipher_len);
+
+        (
+            Packet {
+                header: Header {
+                    ifac_flag: IfacFlag::Open,
+                    header_type: HeaderType::Type1,
+                    context_flag: ContextFlag::Unset,
+                    propagation_type: PropagationType::Broadcast,
+                    destination_type: DestinationType::Link,
+                    packet_type: PacketType::Data,
+                    hops: 0,
+                },
+                ifac: None,
+                destination: *guard.id(),
+                transport: None,
+                context,
+                data: packet_data,
+            },
+            ingress_iface,
+        )
+    };
+
+    let request_id = if context == PacketContext::Request {
+        let hash = packet.hash().to_bytes();
+        let mut request_id = [0u8; 16];
+        request_id.copy_from_slice(&hash[..16]);
+        Some(request_id)
+    } else {
+        None
+    };
+
+    transport.send_direct(ingress_iface, packet).await;
+    Ok(request_id)
+}
+
+pub(super) async fn wait_for_link_request_response(
+    data_rx: &mut tokio::sync::broadcast::Receiver<rns_transport::transport::ReceivedData>,
+    resource_rx: &mut tokio::sync::broadcast::Receiver<rns_transport::resource::ResourceEvent>,
+    expected_destination: AddressHash,
+    expected_link_id: AddressHash,
+    request_id: [u8; 16],
+    timeout: Duration,
+) -> Result<rmpv::Value, String> {
+    let deadline = tokio::time::Instant::now() + timeout;
+    loop {
+        let now = tokio::time::Instant::now();
+        if now >= deadline {
+            return Err("propagation control response timed out".to_string());
+        }
+        let remaining = deadline.saturating_duration_since(now);
+
+        tokio::select! {
+            _ = tokio::time::sleep(remaining) => {
+                return Err("propagation control response timed out".to_string());
+            }
+            result = data_rx.recv() => {
+                match result {
+                    Ok(event) => {
+                        if event.destination != expected_link_id
+                            && event.destination != expected_destination
+                        {
+                            continue;
+                        }
+                        if let Some((response_id, payload)) =
+                            parse_link_response_frame(event.data.as_slice())
+                        {
+                            if response_id == request_id {
+                                return Ok(payload);
+                            }
+                        }
+                    }
+                    Err(tokio::sync::broadcast::error::RecvError::Lagged(_)) => continue,
+                    Err(tokio::sync::broadcast::error::RecvError::Closed) => {
+                        return Err("propagation control response channel closed".to_string());
+                    }
+                }
+            }
+            result = resource_rx.recv() => {
+                match result {
+                    Ok(event) => {
+                        let rns_transport::resource::ResourceEventKind::Complete(complete) =
+                            event.kind
+                        else {
+                            continue;
+                        };
+                        if event.link_id != expected_link_id {
+                            continue;
+                        }
+                        if let Some((response_id, payload)) =
+                            parse_link_response_frame(complete.data.as_slice())
+                        {
+                            if response_id == request_id {
+                                return Ok(payload);
+                            }
+                        }
+                    }
+                    Err(tokio::sync::broadcast::error::RecvError::Lagged(_)) => continue,
+                    Err(tokio::sync::broadcast::error::RecvError::Closed) => {
+                        return Err("propagation control resource channel closed".to_string());
+                    }
+                }
+            }
+        }
+    }
+}
+
+fn parse_link_response_frame(bytes: &[u8]) -> Option<([u8; 16], rmpv::Value)> {
+    let value = rmp_serde::from_slice::<rmpv::Value>(bytes).ok()?;
+    let rmpv::Value::Array(entries) = value else {
+        return None;
+    };
+    if entries.len() != 2 {
+        return None;
+    }
+    let request_bytes = value_to_bytes(entries.first()?)?;
+    if request_bytes.len() != 16 {
+        return None;
+    }
+    let mut request_id = [0u8; 16];
+    request_id.copy_from_slice(request_bytes.as_slice());
+    Some((request_id, entries.get(1)?.clone()))
+}
+
+fn value_to_bytes(value: &rmpv::Value) -> Option<Vec<u8>> {
+    match value {
+        rmpv::Value::Binary(bytes) => Some(bytes.clone()),
+        rmpv::Value::String(text) => {
+            let value = text.as_str()?;
+            if let Ok(decoded) = hex::decode(value) {
+                return Some(decoded);
+            }
+            Some(value.as_bytes().to_vec())
+        }
+        _ => None,
+    }
+}
